Merging conditional jumps
Conditional and unconditional jumps are now only two types of binary
expressions. An unconditional jumps ignore its second child (for
convenient similarity with conditional jumps). A conditional jump
takes an expression child, which is expected to be either 0 or 1. It is
initially generated as being a E_TEST, i.e. the result of a CMP or TEST
instruction stored in the flag register. This approach allows reduction
of conditional expressions.