Fixed a misnaming in postproc_aux1() and improved E_LEA simplification
The variable name for the new right member of the E_MOV instruction to
replace E_LEA was the same as the original instruction, resulting in a
non-replacement. The original right member was then deleted, leading to
undefined behaviour.

An instruction such as 'lea (x,x,n), y' is now replaced for 'y = (n+1)*x'.