Functions are now represented as function_t instances.
function_t groups function address, name (to be found in symbol table
or generated), and its expression, when appliable (i.e. when it's a
function defined in the .text section). Being able to handle local
and external functions the same way is important for later fingerprint
analysis. However, the two should be separated for name solving and in
the list of the interface.  # Please enter the commit message for your
changes. Lines starting # with '#' will be ignored, and an empty message
aborts the commit.