Added back PLT and automatic function naming
post_funs() now extract a sorted lists of all detected functions where
each appears exactly once. They are then named by PLT solving and
automatic naming. Local names solving is to be implemented.

E_CALL and E_JMP to a function currently use a E_IM expression with a
symbol. It is to be replaced by a specific function expression with a
pointer to the function symbol.  # Please enter the commit message for
your changes. Lines starting