Detecting non-returning functions
Typical non-returning functions are exit() and abort(). Functions
which ends up on a non-returning function are non-returning functions
(e.g. usage()). The non-returning functions are for now detected by using
a 10 time iteration on doesReturn(), which checks if the function ends
up on an already known non-returning function. However, using a function
graph should lead to a better way to do this.

The point of detecting non-returning functions is to avoid wrong 'next'
links in blocks. They are removed in breakNonReturning(). Reordering
the way hierarchy is built and function are detected could help breaking
post_funs() in several chunks.